tests: arch: arm: move test arm_irq_vector_table to new ztest API Migrate the testsuite tests/arch/arm/arm_irq_vector_table to the new ztest API.
