+++ tests/arch/arm/arm_irq_vector_table/prj.conf
@@ -1,5 +1,6 @@
 CONFIG_ZTEST=y
 CONFIG_GEN_ISR_TABLES=n
 CONFIG_NUM_IRQS=3
+CONFIG_ZTEST_NEW_API=y
 # Force Bluetooth disable (required by platforms that enable Bluetooth by default)
 CONFIG_BT=n
